Add XML validation and parsing functions
This commit adds two new functions, ValidateXML and ParseXML, to the utils/xml.ts file. The ValidateXML function uses the fast-xml-parser library to validate XML data and returns a boolean indicating whether the data is valid or not. The ParseXML function uses the same library to parse XML data and returns an object representing the parsed data. These functions will be useful for handling XML data in the application.