feat: Add ByteManager, ByteReader, and ByteWriter classes for manipulating binary data
The commit adds three new classes, ByteManager, ByteReader, and ByteWriter, to the utils/bytemgr.ts file. The ByteManager class provides basic functionality for manipulating binary data, such as getting the number of remaining bytes, skipping bytes, and checking if the end of the buffer has been reached. The ByteReader class extends ByteManager and adds methods for reading various types of data from the buffer. The ByteWriter class also extends ByteManager and adds methods for writing various types of data to the buffer. These classes will be useful for handling binary data in the application.